refactor(headless): merge Rust test runner
