+++ .github/headless-rust-test-runner.instructions.md
@@ -1,15 +1,15 @@
 ---
-applyTo: "packages/lynx/headless-rust-test-runner/**"
+applyTo: "packages/lynx/headless-rust-test-runner/**,packages/lynx/engine-bridge/lynx/**"
 ---
 
-`packages/lynx/headless-rust-test-runner` is a Rust-only integration runner for the engine bridge. Keep it as a workspace rlib/bin crate that uses `packages/lynx/engine-bridge/lynx`; do not add JS package metadata around it.
+Keep `lynx-headless-rust-test-runner` as the single Rust-only headless testing crate built on `packages/lynx/engine-bridge/lynx`. Expose Tokio futures and Puppeteer-style page APIs with CDP behavior parity to `packages/testing-library/kitten-lynx` where the headless host supports it. Use the software renderer frame as the screenshot source and the local desktop DebugRouter instead of ADB; do not create a parallel Kitten Lynx Rust crate.
 
-The React fixture bundle is a generated build output, not a copied runner fixture. Build it with `NODE_ENV=production node packages/rspeedy/core/bin/rspeedy.js build --root packages/genui/ui-judge/tests/fixtures/react`, then read `packages/genui/ui-judge/tests/fixtures/react/.generated/main.lynx.bundle` directly from the Rust runner. Keep `fixtures/react/lynx_core.js` aligned with the external Lynx core output used by the runtime. Do not run ESLint, Biome, or dprint over runner fixture files.
+Reuse the React fixture generated under `packages/genui/ui-judge/tests/fixtures/react/.generated`; do not copy generated bundle output into this crate. Runtime-backed CI tests are Linux-only unless the platform contract is deliberately expanded. macOS may keep an ignored diagnostic test.
 
-Runtime-backed tests for this runner are Linux-only unless a PR explicitly expands the platform matrix. Native runtime loading currently expects `lynx_core.js` beside the test executable on Linux; install or copy the fixture core there before creating the view. Assert the rendered software frame through stable visual signals from the fixture source rather than through a fragile golden image.
+The shipped runtime starts its local DebugRouter after app metadata is registered through `lynx_env_set_devtool_app_info` and devtool support is enabled. Scan ports 8901 through 8910 with the Peertalk Initialize request. Unlike the Android transport, the local desktop DebugRouter requires Initialize, ListSession, and CDP messages to reuse the same TCP connection; serialize requests on that initialized stream.
 
-Keep macOS support Rust-only. Do not add Swift, Objective-C, or Objective-C++ sources for the runner. On macOS, let the shipped runtime initialize its own UI message loop before creating the view; do not pre-register the windowless global UI task runner, because that replaces the UI runner with a delegate-only runner and prevents the default ALL_ON_UI TASM actor from executing.
+Keep the renderer and global UI task queues pumping while waiting for frames or interaction updates. On macOS, preserve the Rust-only fake display-link setup used by `headless-rust-test-runner`; on non-macOS platforms, register the windowless global UI task runner once per process.
 
-Inline `data:image/...;base64,...` images are decoded and painted by Clay itself. Do not add `generic_resource_fetcher` plumbing for inline images; use a resource fetcher only when the fixture starts depending on host-provided URLs, fonts, lazy bundles, or template URLs.
+Use Tokio futures for DebugRouter I/O, CDP waits, resource reads, and public waiting APIs. Drive the native task pump with `tokio::select!` while CDP is pending; do not reintroduce an idle callback or a helper reader thread.
 
-For `kRendererTypeSoftware`, keep Clay image resources on the software path before loading templates. The Rust bridge should call a native shim such as `lynx_rust_view_set_use_texture_backend(view, false)` after creating the headless view; otherwise inline images can decode successfully but miss the final software present buffer.
+Dispatch `ElementNode::tap` directly to the Lynx node id with the native `tap` event. Do not derive an absolute point from `DOM.getBoxModel` and do not add coordinate-based tap APIs, because overlay and stacking relationships can make coordinate hit-testing select a different node.
