+++ Cargo.toml
@@ -4,7 +4,6 @@ resolver = "2"
 members = [
   "packages/lynx/engine-bridge/lynx",
   "packages/lynx/headless-rust-test-runner",
-  "packages/lynx/kitten-lynx-rust",
   "packages/react/transform",
   "packages/react/transform/crates/*",
   "packages/react/transform/swc-plugin-reactlynx",
