+++ packages/lynx/headless-rust-test-runner/Cargo.toml
@@ -7,7 +7,12 @@ publish = false
 [dependencies]
 lynx = { path = "../engine-bridge/lynx" }
 png = "0.18.1"
+serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
+thiserror = "2.0.9"
+tokio = { workspace = true, features = ["fs", "io-util", "macros", "net", "rt", "sync", "time"] }
+ureq = "2.7.1"
+url = "2.5.4"
 
 [[bin]]
 name = "lynx-headless-rust-test-runner"
