--- packages/lynx/kitten-lynx-rust/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "kitten-lynx-rust"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-lynx = { path = "../engine-bridge/lynx" }
-png = "0.18.1"
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-thiserror = "2.0.9"
-tokio = { workspace = true, features = ["fs", "io-util", "macros", "net", "rt", "sync", "time"] }
-ureq = "2.7.1"
-url = "2.5.4"
